requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Speech Recognition
SpeechRecognition>=3.10.0

# OpenAI API
openai>=1.0.0

# Wake Word Detection (Picovoice)
pvporcupine>=3.0.0
pyaudio>=0.2.13

# Voice Recognition/Verification
resemblyzer>=0.1.1

# Audio Processing
sounddevice>=0.4.6
soundfile>=0.12.1

# Numerical Computing
numpy>=1.24.0

